Conflicts:
	install_dependencies.sh
	share/make-generic
	share/src/MT/ors_ode.cpp
	share/src/biros/biros.h
	share/src/biros/logging.cpp
	share/src/biros/logging.h
	share/src/motion/GeometricState.cpp
	share/src/motion/MotionController.cpp
	share/src/perception/Makefile
	share/src/relational/robotManipulationInterface.cpp
	share/src/relational/robotManipulationSimulator.cpp
	share/test/hardware_arm/Makefile
	share/test/motion_joystick/main.cpp
	share/test/ors/main.cpp

+++ share/src/relational/plan.cpp
@@ -412,7 +412,7 @@ void NID_UCT::runEpisode(double& value, const SymbolicState& s, uint t) {
           UCB(i) = -11.;
         }
         else
-          UCB(i) = s_a_info->getQvalue(i)  +  c * sqrt(log(s_a_info->getVisits()) / (1.0 * s_a_info->getVisits(i)));
+          UCB(i) = s_a_info->getQvalue(i)  +  c * sqrt(log((double)s_a_info->getVisits()) / (1.0 * s_a_info->getVisits(i)));
       }
       else {
         rules.append(Rule::getDoNothingRule()); // just as a hack
